profile/body 테이블 weight 컬럼의 type을 REAL로 변경 스키마 변경으로 디비 마이그레이션 추가 눈바디 position 입력값으로 이동 기능 추가 summaryFragment에서 diet/workout find 값이 없을 경우에도 화면에 이전 값이 보이는 문제 수정

+++ app/src/main/java/com/seedit/diet/MyInfoActivity.kt
@@ -47,7 +47,7 @@ class MyInfoActivity : AppCompatActivity() {
             CLog.i("화면 갱신 profile data size=${it?.size?:0}")
             it?.let {
                 if(it.isEmpty())
-                    viewModel.insert(if(BuildConfig.DEBUG) ProfileEntity(0,null,"test",0,Calendar.getInstance().apply { set(1980,10,17) }.time,80,170,60,400,300,2000,Calendar.getInstance().apply { set(2018,10,17) }.time) else ProfileEntity(0))
+                    viewModel.insert(if(BuildConfig.DEBUG) ProfileEntity(0,null,"test",0,Calendar.getInstance().apply { set(1980,10,17) }.time,80f,170,60,400,300,2000,Calendar.getInstance().apply { set(2018,10,17) }.time) else ProfileEntity(0))
                     //viewModel.insert(ProfileEntity(0))
                 else
                 //if(it.isNotEmpty())
@@ -161,7 +161,7 @@ class MyInfoActivity : AppCompatActivity() {
         {
             override fun afterTextChanged(s: Editable?) {
                 try {
-                    val weight = editWeight.text.toString().toInt()
+                    val weight = editWeight.text.toString().toFloat()
                     val height = editHeight.text.toString().toInt()
 
                     viewModel.getProfile()?.weight=weight
@@ -276,7 +276,7 @@ class MyInfoActivity : AppCompatActivity() {
             showDlg(dlg,"몸무게를",editWeight)
             return
         }
-        viewModel.getProfile()?.weight=editWeight.text.toString().toInt()
+        viewModel.getProfile()?.weight=editWeight.text.toString().toFloat()
         if(editHeight.text.isEmpty())
         {
             showDlg(dlg,"키를",editHeight)
